requirements_dev.txt
alembic==1.13.1
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astroid==3.0.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
dill==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
flask==3.0.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-PyMongo==2.3.0
flask-wtf==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
mypy==1.7.1
mypy-extensions==1.0.0
ngrok==0.12.1
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg==3.1.16
psycopg-binary==3.1.16
pycparser==2.21
pylint==3.0.3
pymongo==4.6.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
tomli==2.0.1
tomlkit==0.12.3
typing-extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
werkzeug==3.0.1
wtforms==3.1.1
yarl==1.9.4
zipp==3.17.0
